Simplifying tests for wasm
