+++ .github/workflows/npm-publish.yml
@@ -19,23 +19,36 @@ jobs:
       fail-fast: false
 
     steps:
-    - name: Run wasm tests
-      if: startsWith(matrix.os, 'ubuntu') && startsWith(matrix.python, '3.7')
-      run: |
-        rustup target add wasm32-unknown-unknown
-        cargo install wasm-bindgen-cli --version 0.2.80
-        cd mock-test && npm install && make && ../resources/tests/linux-install-node.sh ./build/runmock.js
+      - uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
 
-    - name: Test for secrets access
-      id: check_secrets
-      shell: bash
-      run: |
-        unset HAS_SECRET
-        if [ -n "$SECRET" ]; then HAS_SECRET='true' ; fi
-        echo ::set-output name=HAS_SECRET::${HAS_SECRET}
-      env:
-        SECRET: "${{ secrets.test_pypi_password }}"
+      - name: Set up rusts
+        uses: actions-rs/toolchain@v1
+        with:
+          toolchain: stable
+          components: rustfmt, clippy
 
-    - name: Publish wasm
-      if: steps.check_secrets.HAS_SECRET
-      run: cd pkg && npm publish --access public
+      - name: install wasm-pack
+        run: cargo install wasm-pack
+
+      - name: wasm-pack build and pack
+        run: wasm-pack build --release --target=nodejs wasm && wasm-pack pack wasm
+
+      - name: Setup Node 16.x
+        uses: actions/setup-node@v3
+        with:
+          node-version: '16.x'
+
+      - name: Test wasm
+        run: node tests/index.js
+
+      - name: Test for secrets access
+        id: check_secrets
+        shell: bash
+        run: |
+          unset HAS_SECRET
+          if [ -n "$SECRET" ]; then HAS_SECRET='true' ; fi
+          echo ::set-output name=HAS_SECRET::${HAS_SECRET}
+        env:
+          SECRET: "${{ secrets.test_pypi_password }}"
