--- mock-test/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "clvm_tools_rs_mock_test",
-  "scripts": {
-    "build": "webpack"
-  },
-  "dependencies": {
-    "binascii": "^0.0.2",
-    "browserify": "^17.0.0",
-    "rollup": "^2.77.2",
-    "webpack": "^5.72.0",
-    "webpack-cli": "^4.10.0"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.1.6",
-    "@babel/node": "^7.10.5",
-    "@babel/preset-env": "^7.1.6",
-    "babel-loader": "^8.0.4",
-    "babelify": "^10.0.0"
-  }
-}
