+++ wasm/tests/test-data/fact.sym
