+++ wasm/tests/test-data/pwcoin.cl
